hcom/requirements.txt
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
Django==3.2.4
flake8==3.9.2
fontawesome-free==5.15.4
isort==5.9.1
jedi==0.17.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
nodeenv==1.6.0
parso==0.7.1
pluggy==0.13.1
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
pylint==2.8.3
pyright==0.0.1
python-jsonrpc-server==0.4.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.0.1
pytz==2021.1
rope==0.19.0
snowballstemmer==2.1.0
sqlparse==0.4.1
toml==0.10.2
ujson==4.0.2
wrapt==1.12.1
yapf==0.31.0
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
crispy-bootstrap5==0.7
debugpy==1.6.7
Django==3.2.4
django-crispy-forms==2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
flake8==3.9.2
fontawesome-free==5.15.4
isort==5.9.1
jedi==0.17.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
nodeenv==1.6.0
parso==0.7.1
Pillow==9.5.0
pluggy==0.13.1
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
pylint==2.8.3
pyright==0.0.1
python-jsonrpc-server==0.4.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.0.1
pytz==2021.1
rope==0.19.0
snowballstemmer==2.1.0
sqlparse==0.4.1
toml==0.10.2
ujson==4.0.2
wrapt==1.12.1
yapf==0.31.0
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
crispy-bootstrap5==0.7
debugpy==1.6.7
Django==3.2.4
django-crispy-forms==2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
flake8==3.9.2
fontawesome-free==5.15.4
isort==5.9.1
jedi==0.17.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
nodeenv==1.6.0
parso==0.7.1
Pillow==9.5.0
pluggy==0.13.1
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
pylint==2.8.3
pyright==0.0.1
python-jsonrpc-server==0.4.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.0.1
pytz==2021.1
rope==0.19.0
snowballstemmer==2.1.0
sqlparse==0.4.1
toml==0.10.2
ujson==4.0.2
wrapt==1.12.1
yapf==0.31.0
asgiref==3.7.2
astroid==2.5.6
autopep8==1.5.7
crispy-bootstrap5==0.7
debugpy==1.6.7
Django==4.2.6
django-crispy-forms==2.0
django-rest-framework==0.1.0
django-stubs==4.2.4
django-stubs-ext==4.2.2
djangorestframework==3.14.0
djangorestframework-datatables==0.7.0
flake8==3.9.2
fontawesome-free==5.15.4
isort==5.9.1
jedi==0.17.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy==1.6.1
mypy-extensions==1.0.0
nodeenv==1.6.0
parso==0.7.1
Pillow==9.5.0
pluggy==0.13.1
pycodestyle==2.7.0
pydocstyle==6.1.1
pyflakes==2.3.1
pylint==2.8.3
pyright==1.1.331
python-jsonrpc-server==0.4.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.0.1
pytz==2021.1
rope==0.19.0
snowballstemmer==2.1.0
sqlparse==0.4.1
toml==0.10.2
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.8.0
ujson==4.0.2
wrapt==1.12.1
yapf==0.31.0
